Extract static/js/pm_list.js.
This handles most of the details of building the Private Messages
section in the upper left corner of the app.
