+++ zproject/settings.py
@@ -772,6 +772,7 @@ JS_SPECS = {
             'js/people.js',
             'js/unread.js',
             'js/topic_list.js',
+            'js/pm_list.js',
             'js/stream_list.js',
             'js/filter.js',
             'js/message_list_view.js',
